pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asr-cpu-benchmark"
version = "0.1.0"
description = "Whisper CPU acceleration benchmarks"
requires-python = ">=3.10"
dependencies = [
    "jiwer>=3.0",
    "psutil>=5.9",
    "tabulate>=0.9",
    "datasets>=2.14",
    "soundfile>=0.12",
    "librosa>=0.10",
    "pyyaml>=6.0",
    "torchcodec>=0.11",
]

[project.optional-dependencies]
openai-whisper = ["openai-whisper"]
faster-whisper = ["faster-whisper"]
onnx = ["optimum[onnxruntime]", "transformers"]
torch-opt = ["torch", "transformers"]
all = [
    "openai-whisper",
    "faster-whisper",
    "optimum[onnxruntime]",
    "transformers",
    "torch",
]

[tool.setuptools.packages.find]
include = ["benchmark*"]
